feat: move branch dialog to playground
